Split CodeSet type into trait and impls
